Change tooltip adjust method to flip
This makes more efficient use of the available screen estate and
greatly improves navigation for deeply nested tooltips.

Only disadvantage is that the tooltip can extend outside the document in
edge cases with very limited vertial space (<200px). As these
generally don't occur in regular use of the app we can disregard
this for now.